moved remaining helpers except for publishHelpers
