+++ server/routes/api/channelAvailability.js
@@ -1,5 +1,5 @@
 const { checkChannelAvailability } = require('../../controllers/publishController.js');
-const { sendGATimingEvent } = require('../../helpers/googleAnalytics.js');
+const { sendGATimingEvent } = require('../../utils/googleAnalytics.js');
 const { handleErrorResponse } = require('../utils/errorHandlers.js');
 
 /*
