+++ server/routes/api/claimAvailability.js
@@ -1,5 +1,5 @@
 const { claimNameIsAvailable } = require('../../controllers/publishController.js');
-const { sendGATimingEvent } = require('../../helpers/googleAnalytics.js');
+const { sendGATimingEvent } = require('../../utils/googleAnalytics.js');
 const { handleErrorResponse } = require('../utils/errorHandlers.js');
 
 /*
