+++ server/routes/api/claimGet.js
@@ -1,4 +1,4 @@
-const { getClaim } = require('../../helpers/lbryApi.js');
+const { getClaim } = require('../../utils/lbryApi.js');
 const { addGetResultsToFileData, createFileData } = require('../../helpers/publishHelpers.js');
 const { handleErrorResponse } = require('../utils/errorHandlers.js');
 const db = require('../../models');
