+++ server/routes/utils/multipartMiddleware.js
@@ -1,5 +1,5 @@
 const multipart = require('connect-multiparty');
-const { publishing: { uploadDirectory } } = require('../../config/siteConfig.js');
+const { publishing: { uploadDirectory } } = require('../../../config/siteConfig.js');
 const multipartMiddleware = multipart({uploadDir: uploadDirectory});
 
 module.exports = multipartMiddleware;
